fixed the utils issue by moving all the scripts/files under the same terminology directory
